Bug fixes in decoding.

+++ proton-j/src/main/java/org/apache/qpid/proton/transport2/Source.java
@@ -158,12 +158,12 @@ public final class Source extends Terminus implements Encodable
                 Boolean dynamic = (Boolean) l.get(4);
                 source.setDynamic(dynamic == null ? false : dynamic);
             case 7:
-                source.setTimeout(l.get(3) == null ? 0 : (Long) l.get(3));
+                source.setTimeout(l.get(3) == null ? 0 : (int) l.get(3));
             case 8:
                 source.setExpiryPolicy(l.get(2) == null ? TerminusExpiryPolicy.SESSION_END : TerminusExpiryPolicy
                         .getEnum((String) l.get(2)));
             case 9:
-                source.setDurable(l.get(1) == null ? TerminusDurability.NONE : TerminusDurability.get((Byte) l.get(1)));
+                source.setDurable(l.get(1) == null ? TerminusDurability.NONE : TerminusDurability.get(((Integer)l.get(1)).byteValue()));
             case 10:
                 source.setAddress((String) l.get(0));
             }
